refactor(pr12): move src/scene/ into src/documents/scene/ + camelCase subfolders (G6b.1-G6b.9)
Completes the Foundry-alignment story by collapsing the top-level scene/ folder into documents/scene/ (Scene, TokenDocument, RegionDocument, RegionBehaviour are all Foundry documents). Also normalizes the kebab-case subfolder names to camelCase: region-document -> regionDocument, token-document -> tokenDocument, region-behaviour -> regionBehaviour (British spelling preserved per Foundry).

Path alias @scene/* retargeted in tsconfig.json, vite.config.ts, vitest.config.ts. Two-step renames used for kebab->camel on Windows to avoid case-collision. Workspace string replaces sweep @scene/region-document/ -> @scene/regionDocument/ (and the other two pairs); global.mts relative ./scene/ imports fixed by hand.
