+++ src/canvas/RegionDnd35e.mts
@@ -1,4 +1,4 @@
-import { RegionDocumentDnd35e } from '@scene/region-document/RegionDocumentDnd35e.mjs';
+import { RegionDocumentDnd35e } from '@scene/regionDocument/RegionDocumentDnd35e.mjs';
 
 class RegionDnd35e<TDocument extends RegionDocumentDnd35e = RegionDocumentDnd35e> extends fc.placeables.Region<TDocument> {
