+++ src/canvas/token/TokenDnd35e.mts
@@ -1,5 +1,5 @@
 import type TokenLayer from '@client/canvas/layers/tokens.mjs';
-import type { TokenDocumentDnd35e } from '@scene/token-document/TokenDocumentDnd35e.mjs';
+import type { TokenDocumentDnd35e } from '@scene/tokenDocument/TokenDocumentDnd35e.mjs';
 
 class TokenDnd35e<TDocument extends TokenDocumentDnd35e = TokenDocumentDnd35e> extends fc.placeables.Token<TDocument> {
   declare readonly layer: TokenLayer<this>;
