+++ src/documents/scene/index.mts
@@ -1,0 +1,17 @@
+import { DIFFICULT_TERRAIN_GRADES,EnvironmentBehaviorType, EnvironmentFeatureBehaviorType, RegionBehaviorDnd35e } from './regionBehaviour/index.mjs';
+import { RegionDocumentDnd35e } from './regionDocument/RegionDocumentDnd35e.mjs';
+import { SceneDnd35e } from './SceneDnd35e.mjs';
+import type { TokenDocumentDnd35e } from './tokenDocument/index.mjs';
+
+export {
+  DIFFICULT_TERRAIN_GRADES,
+  EnvironmentBehaviorType,
+  EnvironmentFeatureBehaviorType,
+  RegionBehaviorDnd35e,
+  RegionDocumentDnd35e,
+  SceneDnd35e,
+};
+
+export type {
+  TokenDocumentDnd35e,
+};
