+++ src/documents/scene/regionDocument/RegionDocumentDnd35e.mts
@@ -1,6 +1,6 @@
 import type EmbeddedCollection from '@common/abstract/embedded-collection.mjs';
 
-import type { SpecificRegionBehavior } from '../region-behaviour/types.mjs';
+import type { SpecificRegionBehavior } from '../regionBehaviour/types.mjs';
 import { SceneDnd35e } from '../SceneDnd35e.mjs';
 // import type EmbeddedCollection from "@common/abstract/embedded-collection.d.mts";
