--- src/scene/index.mts
@@ -1,17 +1,0 @@
-import { DIFFICULT_TERRAIN_GRADES,EnvironmentBehaviorType, EnvironmentFeatureBehaviorType, RegionBehaviorDnd35e } from './region-behaviour/index.mjs';
-import { RegionDocumentDnd35e } from './region-document/RegionDocumentDnd35e.mjs';
-import { SceneDnd35e } from './SceneDnd35e.mjs';
-import type { TokenDocumentDnd35e } from './token-document/index.mjs';
-
-export {
-  DIFFICULT_TERRAIN_GRADES,
-  EnvironmentBehaviorType,
-  EnvironmentFeatureBehaviorType,
-  RegionBehaviorDnd35e,
-  RegionDocumentDnd35e,
-  SceneDnd35e,
-};
-
-export type {
-  TokenDocumentDnd35e,
-};
